+++ include/mruby_integration/models/render_texture.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "mruby.h"
+#include "raylib.h"
+
+extern RClass *RenderTexture_class;
+
+void append_models_RenderTexture(mrb_state*);