+++ include/mruby_integration/struct_types.hpp
@@ -16,5 +16,6 @@ extern mrb_data_type Image_type;
 extern mrb_data_type Music_type;
 extern mrb_data_type Rectangle_type;
 extern mrb_data_type Sound_type;
+extern mrb_data_type RenderTexture_type;
 extern mrb_data_type Texture2D_type;
 extern mrb_data_type Vector2_type;