+++ src/mruby_integration/structs.cpp
@@ -5,6 +5,7 @@
 #include "mruby_integration/models/music.hpp"
 #include "mruby_integration/models/rectangle.hpp"
 #include "mruby_integration/models/sound.hpp"
+#include "mruby_integration/models/render_texture.hpp"
 #include "mruby_integration/models/texture2d.hpp"
 #include "mruby_integration/models/vector2.hpp"
 
@@ -16,6 +17,7 @@ void append_structs(mrb_state *mrb) {
   append_models_Music(mrb);
   append_models_Rectangle(mrb);
   append_models_Sound(mrb);
+  append_models_RenderTexture(mrb);
   append_models_Texture2D(mrb);
   append_models_Vector2(mrb);
 }